Moved langium-grammar-hover-provider to src/grammar/lsp
